backend/requirements.txt
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# sqlalchemy[asyncio]==2.0.23
# asyncpg==0.29.0
# alembic==1.12.1
# pydantic==2.5.0
# pydantic-settings==2.0.3
# python-multipart==0.0.6
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
# python-dotenv==1.0.0
# websockets==12.0
# langgraph==0.1.19
# langchain==0.1.0
# langchain-google-genai==1.0.10
# google-generativeai==0.8.3
# redis==5.0.1
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.25.2


# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# WebSockets
websockets==12.0

# LangGraph and LangChain - Compatible versions
langgraph==0.2.16
langchain==0.2.14
langchain-core==0.2.38
langchain-google-genai==1.0.10
google-generativeai==0.7.2

# Redis
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2